refactor(cli): migrate `sync`
